Moves IVisabFile and VISABFileBase to repository package
