+++ src/main/java/org/visab/repository/IVISABFile.java
@@ -1,4 +1,4 @@
-package org.visab.processing;
+package org.visab.repository;
 
 import java.sql.Timestamp;
